pyproject.toml
[tool.pdm]

[project]
name = "botty"
version = "0.1.0"
description = ""
authors = [
    {name = "levch", email = "levchenko.d.a1998@gmail.com"},
]
dependencies = [
    "python-telegram-bot>=20.1",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-pep517>=1.0"]
build-backend = "pdm.pep517.api"
